return rel (current table) from functions instead of keeping in context
